issue 24: move event handler into its own package
